# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ApplicationInputAidType = typeof ApplicationInputAidType[keyof typeof ApplicationInputAidType];


export const ApplicationInputAidType = {
  emergency_aid: 'emergency_aid',
  education_grant: 'education_grant',
  business_grant: 'business_grant',
  housing_assistance: 'housing_assistance',
  medical_aid: 'medical_aid',
  other: 'other',
} as const;

export interface ApplicationInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 3 */
  email: string;
  phone: string;
  address: string;
  aidType: ApplicationInputAidType;
  requestedAmount: number;
  /** @minLength 1 */
  purpose: string;
  /** @minLength 10 */
  description: string;
  /** @nullable */
  householdSize?: number | null;
  /** @nullable */
  annualIncome?: number | null;
}

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  pending: 'pending',
  under_review: 'under_review',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Application {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  address: string;
  aidType: string;
  requestedAmount: number;
  purpose: string;
  description: string;
  /** @nullable */
  householdSize?: number | null;
  /** @nullable */
  annualIncome?: number | null;
  status: ApplicationStatus;
  submittedAt: string;
}

export interface Testimonial {
  id: number;
  name: string;
  location: string;
  aidType: string;
  message: string;
  rating: number;
  /** @nullable */
  avatarInitials?: string | null;
  createdAt: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @minLength 1 */
  subject: string;
  /** @minLength 10 */
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  createdAt: string;
}

export interface OrgStats {
  totalApplications: number;
  approvedApplications: number;
  totalDisbursed: number;
  totalTestimonials: number;
}

export interface ErrorResponse {
  error: string;
}