# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: applications
    description: Grant/Aid application operations
  - name: testimonials
    description: Testimonial operations
  - name: contact
    description: Contact form operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /applications:
    post:
      operationId: submitApplication
      tags: [applications]
      summary: Submit a grant or aid application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List all approved testimonials
      responses:
        "200":
          description: List of testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Contact message submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [applications]
      summary: Get organization statistics
      responses:
        "200":
          description: Organization statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApplicationInput:
      type: object
      required:
        - firstName
        - lastName
        - email
        - phone
        - address
        - aidType
        - requestedAmount
        - purpose
        - description
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 3
        phone:
          type: string
        address:
          type: string
        aidType:
          type: string
          enum: [emergency_aid, education_grant, business_grant, housing_assistance, medical_aid, other]
        requestedAmount:
          type: number
        purpose:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 10
        householdSize:
          type: ["number", "null"]
        annualIncome:
          type: ["number", "null"]

    Application:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
        - phone
        - address
        - aidType
        - requestedAmount
        - purpose
        - description
        - status
        - submittedAt
      properties:
        id:
          type: number
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        aidType:
          type: string
        requestedAmount:
          type: number
        purpose:
          type: string
        description:
          type: string
        householdSize:
          type: ["number", "null"]
        annualIncome:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, under_review, approved, rejected]
        submittedAt:
          type: string

    Testimonial:
      type: object
      required:
        - id
        - name
        - location
        - aidType
        - message
        - rating
        - createdAt
      properties:
        id:
          type: number
        name:
          type: string
        location:
          type: string
        aidType:
          type: string
        message:
          type: string
        rating:
          type: number
        avatarInitials:
          type: ["string", "null"]
        createdAt:
          type: string

    ContactInput:
      type: object
      required:
        - name
        - email
        - subject
        - message
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 3
        phone:
          type: ["string", "null"]
        subject:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 10

    ContactMessage:
      type: object
      required:
        - id
        - name
        - email
        - subject
        - message
        - createdAt
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string
        createdAt:
          type: string

    OrgStats:
      type: object
      required:
        - totalApplications
        - approvedApplications
        - totalDisbursed
        - totalTestimonials
      properties:
        totalApplications:
          type: number
        approvedApplications:
          type: number
        totalDisbursed:
          type: number
        totalTestimonials:
          type: number

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string