# lib/api-zod/src/generated/types/application.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationStatus } from './applicationStatus';

export interface Application {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  address: string;
  aidType: string;
  requestedAmount: number;
  purpose: string;
  description: string;
  /** @nullable */
  householdSize?: number | null;
  /** @nullable */
  annualIncome?: number | null;
  status: ApplicationStatus;
  submittedAt: string;
}