# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationInputAidType } from './applicationInputAidType';

export interface ApplicationInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 3 */
  email: string;
  phone: string;
  address: string;
  aidType: ApplicationInputAidType;
  requestedAmount: number;
  /** @minLength 1 */
  purpose: string;
  /** @minLength 10 */
  description: string;
  /** @nullable */
  householdSize?: number | null;
  /** @nullable */
  annualIncome?: number | null;
}