# lib/api-zod/src/generated/types/contactMessage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  createdAt: string;
}