# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  name: string;
  location: string;
  aidType: string;
  message: string;
  rating: number;
  /** @nullable */
  avatarInitials?: string | null;
  createdAt: string;
}